requirements.txt
streamlit>=1.35.0,<2.0.0
folium>=0.17.0,<1.0.0
streamlit-folium>=0.22.0,<1.0.0
google-cloud-bigquery>=3.20.0,<4.0.0
google-cloud-aiplatform>=1.55.0,<2.0.0
google-adk>=1.0.0,<2.0.0
google-genai>=1.0.0
